backend/daktariml/requirements.txt
django-cors-headers
librosa
resampy
djangorestframework
djangorestframework-simplejwt
pillow
certifi
